backend/requirements.txt
Fabric
Flask-SQLAlchemy
Flask-WTF
WTForms
coverage
gqlaclhemy
# ecdsa
# paramiko
# pycrypto
